+++ src/main/java/pl/asie/computronics/cc/multiperipheral/OpenMultiPeripheral.java
@@ -1,12 +1,16 @@
 package pl.asie.computronics.cc.multiperipheral;
 
+import cpw.mods.fml.common.Optional;
 import dan200.computercraft.api.peripheral.IPeripheral;
+import li.cil.oc.api.network.BlacklistedPeripheral;
 import pl.asie.computronics.api.multiperipheral.WrappedMultiPeripheral;
+import pl.asie.computronics.reference.Mods;
 
 /**
  * @author Vexatos
  */
-public class OpenMultiPeripheral extends WrappedMultiPeripheral {
+@Optional.Interface(iface = "li.cil.oc.api.network.BlacklistedPeripheral", modid = Mods.OpenComputers)
+public class OpenMultiPeripheral extends WrappedMultiPeripheral implements BlacklistedPeripheral {
 	private final boolean derped;
 
 	public OpenMultiPeripheral(IPeripheral peripheral) {
@@ -22,5 +26,11 @@ public class OpenMultiPeripheral extends WrappedMultiPeripheral {
 	@Override
 	public String[] getMethodNames() {
 		return derped ? new String[] { "open_peripherals_derped" } : super.getMethodNames();
+	}
+
+	@Override
+	@Optional.Method(modid = Mods.OpenComputers)
+	public boolean isPeripheralBlacklisted() {
+		return true;
 	}
 }