+++ src/main/java/pl/asie/computronics/integration/enderio/DriverRedstoneControllable.java
@@ -16,7 +16,7 @@ import pl.asie.computronics.integration.CCMultiPeripheral;
 import pl.asie.computronics.integration.ManagedEnvironmentOCTile;
 import pl.asie.computronics.reference.Names;
 
-import java.util.LinkedHashMap;
+import java.util.ArrayList;
 import java.util.Locale;
 
 /**
@@ -38,11 +38,9 @@ public class DriverRedstoneControllable {
 	}
 
 	private static Object[] modes() {
-		LinkedHashMap<Integer, String> modes = new LinkedHashMap<Integer, String>();
-		int i = 1;
+		ArrayList<String> modes = new ArrayList<String>();
 		for(RedstoneControlMode mode : RedstoneControlMode.values()) {
-			modes.put(i, mode.name().toLowerCase(Locale.ENGLISH));
-			i++;
+			modes.add(mode.name().toLowerCase(Locale.ENGLISH));
 		}
 		return new Object[] { modes };
 	}
@@ -117,11 +115,11 @@ public class DriverRedstoneControllable {
 
 		@Override
 		public Object[] callMethod(IComputerAccess computer, ILuaContext context, int method, Object[] arguments) throws LuaException, InterruptedException {
-			switch(method){
-				case 0:{
+			switch(method) {
+				case 0: {
 					return DriverRedstoneControllable.getRedstoneMode(tile);
 				}
-				case 1:{
+				case 1: {
 					if(arguments.length < 1 || !(arguments[0] instanceof String)) {
 						throw new LuaException("first argument needs to be a string");
 					}
@@ -131,7 +129,7 @@ public class DriverRedstoneControllable {
 						throw new LuaException(e.getMessage());
 					}
 				}
-				case 2:{
+				case 2: {
 					return DriverRedstoneControllable.modes();
 				}
 			}