+++ src/main/java/pl/asie/computronics/integration/railcraft/DriverRoutingDetector.java
@@ -21,7 +21,6 @@ import pl.asie.computronics.integration.util.RoutingTableUtil;
 import pl.asie.computronics.reference.Names;
 
 import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -35,14 +34,17 @@ public class DriverRoutingDetector {
 			&& ((DetectorRouting) tile.getDetector()).getInventory().getStackInSlot(0).getItem() instanceof ItemRoutingTable) {
 			if(!((DetectorRouting) tile.getDetector()).isSecure()) {
 				List<List<String>> pages = ItemRoutingTable.getPages(((DetectorRouting) tile.getDetector()).getInventory().getStackInSlot(0));
-				LinkedHashMap<Integer, String> pageMap = new LinkedHashMap<Integer, String>();
-				int i = 1;
+				if(pages == null) {
+					return new Object[] { false, "no valid routing table found"};
+				}
+				ArrayList<String> pageMap = new ArrayList<String>();
+				int i = 0;
 				for(List<String> currentPage : pages) {
 					for(String currentLine : currentPage) {
-						pageMap.put(i, currentLine);
+						pageMap.add(i, currentLine);
 						i++;
 					}
-					pageMap.put(i, "{newpage}");
+					pageMap.add(i, "{newpage}");
 					i++;
 				}
 				if(pageMap.get(i - 1).equals("{newpage}")) {
@@ -183,20 +185,20 @@ public class DriverRoutingDetector {
 		public Object[] callMethod(IComputerAccess computer, ILuaContext context,
 			int method, Object[] arguments) throws LuaException,
 			InterruptedException {
-			switch(method){
-				case 0:{
+			switch(method) {
+				case 0: {
 					return DriverRoutingDetector.getRoutingTable(tile);
 				}
-				case 1:{
+				case 1: {
 					if(arguments.length < 1 || !(arguments[0] instanceof Map)) {
 						throw new LuaException("first argument needs to be a table");
 					}
 					return DriverRoutingDetector.setRoutingTable(tile, arguments);
 				}
-				case 2:{
+				case 2: {
 					return DriverRoutingDetector.getRoutingTableTitle(tile);
 				}
-				case 3:{
+				case 3: {
 					if(arguments.length < 1 || !(arguments[0] instanceof String)) {
 						throw new LuaException("first argument needs to be a string");
 					}